+++ app/src/main/java/com/mironov/coursework/ui/chat/received/ReceivedViewHolder.kt
@@ -7,8 +7,8 @@ import com.mironov.coursework.domain.entity.Message
 
 class ReceivedViewHolder(
     private val binding: ReceivedMessageItemBinding,
-    private val addReaction: (Int) -> Unit,
-    private val onReactionClickListener: (id: Int, emoji: Int) -> Unit,
+    private val addReaction: (Long) -> Unit,
+    private val onReactionClickListener: (id: Long, emoji: Int) -> Unit,
 ) : RecyclerView.ViewHolder(binding.root) {
 
     fun bind(model: Message) {
