+++ app/src/main/java/com/mironov/coursework/ui/chat/sent/SentViewHolder.kt
@@ -6,8 +6,8 @@ import com.mironov.coursework.domain.entity.Message
 
 class SentViewHolder(
     private val binding: SentMessageItemBinding,
-    private val addReaction: (Int) -> Unit,
-    private val onReactionClickListener: (id: Int, emoji: Int) -> Unit,
+    private val addReaction: (Long) -> Unit,
+    private val onReactionClickListener: (id: Long, emoji: Int) -> Unit,
 ) : RecyclerView.ViewHolder(binding.root) {
 
     fun bind(model: Message) {
